Исправил удаление токена и добавил отбрасывание просроченных токенов

+++ models/session.go
@@ -3,8 +3,8 @@ package models
 type Session struct {
 	Id    string `json:"id"`
 	Token string `json:"token"`
-	//time to live сколько сессия еще будет жить
-	TTL int32 `json:"ttl"`
+	//time to live когда сессия умрет
+	TTL int64 `json:"ttl"`
 
 	User *User
 }
